backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
openai
python-dotenv
